feat: route to index page on leave
