+++ components/layout.module.css
@@ -59,7 +59,7 @@
 }
 
 .header-right {
-  margin: var(--space-2x) var(--space-16x);
+  margin: var(--space-2x) var(--space-4x);
   display: none;
 }
 
@@ -114,65 +114,4 @@
 .logo {
   display: flex;
   width: 40px;
-}
-
-.cta-btn {
-  cursor: pointer;
-  font-weight: 500;
-  font-size: 15px;
-  color: white;
-  background: #323332;
-  border: 2px solid #8465cb;
-  border-radius: 8px;
-  padding: 10px 16px;
-  position: relative;
-}
-
-.tooltip {
-  display: none;
-  position: absolute;
-  top: 70px;
-  right: 65px;
-  background: #444443;
-  border-radius: 8px;
-  padding: 12px;
-  font-weight: normal;
-  font-size: 13px;
-  line-height: 20px;
-  width: 250px;
-  justify-content: space-between;
-  gap: 5px;
-  align-items: flex-start;
-}
-
-.overlay {
-  background-color: rgba(0, 0, 0, 0.9);
-  position: fixed;
-  inset: 0;
-}
-
-.content {
-  max-height: 760px;
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  font-family: Inter;
-  background: #111111;
-  border-radius: 8px;
-  padding: 24px;
-  color: white;
-  text-align: center;
-  overflow-y: scroll;
-}
-
-.content p {
-  margin: 0;
-}
-
-.close-btn {
-  width: 100%;
-  color: white;
-  display: flex;
-  justify-content: flex-end;
 }
